Extract `trpcWrapper.tsx`
